flaskapp/requirements.txt
Flask==2.0.2
webargs==8.0.1
google-cloud-storage==1.44.0
gunicorn==20.1.0
pandas==1.3.5